move `package cleantest` to where the interface is
