+++ internal/pkg/cli/svc_delete_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/copilot-cli/internal/pkg/cli/clean/cleantest"
+	"github.com/aws/copilot-cli/internal/pkg/cli/cleantest"
 	"github.com/aws/copilot-cli/internal/pkg/cli/mocks"
 	"github.com/aws/copilot-cli/internal/pkg/config"
 	"github.com/aws/copilot-cli/internal/pkg/manifest/manifestinfo"
